extracted deposit function from evm home context
