+++ src/contexts/NetworkManagerContext/index.ts
@@ -1,4 +1,1 @@
-export {
-  NetworkManagerProvider,
-  useNetworkManager,
-} from "./NetworkManagerContext";
+export * from "./NetworkManagerContext";
